Add create group chats, sort chatlist, cleanup log
